% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_search.R
\name{binary_search}
\alias{binary_search}
\title{binary_search}
\usage{
binary_search(
  target_value,
  monotone_function,
  init_bounds = NULL,
  error_gap = 1e-06,
  max_iters = 100L
)
}
\arguments{
\item{target_value}{The value the function should achieve.}

\item{monotone_function}{The function we want to find the  relevant input for.}

\item{init_bounds}{Default NULL; Bounds on the function being searched.}

\item{error_gap}{Binary search until the gap between the function on the input and the target value is smaller than this number.}

\item{max_iters}{Maximum umber of iterations to try.}
}
\value{
The input that gives \code{target_value} as output.
}
\description{
Generic binary search algo: finding the input.
}
