% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite_matches.R
\name{bipartite_matches}
\alias{bipartite_matches}
\title{bipartite_matches}
\usage{
bipartite_matches(
  dist_mat,
  treat_vec,
  match_method = c("with_replacement", "optimal", "greedy"),
  n_sinks = NULL,
  tol_val = NULL,
  weight_vec = NULL
)
}
\arguments{
\item{dist_mat}{Matrix of pairwise distances.}

\item{treat_vec}{Vector representing all subjects; 0 for control,
1 for treated.}

\item{match_method}{Choice of the three matching methods:
\describe{
  \item{"with_replacement"}{Finds smallest control for each treatment}
  \item{"greedy"}{Greedily generates pairs. Note that
  the order for choosing the greedy pairs is random, which is not the only
  possible solution.}
  \item{"optimal"}{Minimum total distance}
}}

\item{n_sinks}{How many sinks to use; can be vector. Default NULL to match all and ignore sinks.}

\item{tol_val}{Tolerance for solving optimal matches (how far from the true optimal value is acceptable). Speed with large value, accuracy with small.}

\item{weight_vec}{Default \code{NULL}: optionally supply the weight vector used to generate \code{dist_mat}}
}
\value{
A list with five elements and an optional sixth:
  \describe{
    \item{\code{treat_index}}{index of treated units, from all units}
    \item{\code{treat_index_within}}{index of treated units,
          from the set of treated}
    \item{\code{control_index}}{index of control units, from all units}
    \item{\code{control_index_within}}{index of control units,
          from the set of control}
    \item{\code{distance}}{distances between the pairs}
    \item{\code{weight_vec}}{weight vector used to generate
          \code{dist_mat} if supplied}
  }
}
\description{
Generates bipartite matched pairs either with replacement, greedily without replacement, or optimally.
}
