% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_func.R
\name{boot_func}
\alias{boot_func}
\title{boot_func}
\usage{
boot_func(vec, func, resamples = 100L)
}
\arguments{
\item{vec}{Vector to bootstrap resample from.}

\item{func}{Function to call on the samples.}

\item{resamples}{How many resamples to do.}
}
\value{
Returns vector of length \code{resamples}, each from \code{func} applied to a bootstrap resample.
}
\description{
Resampling bootstrap of a vector for a given function.
}
