% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_score.R
\name{brier_score}
\alias{brier_score}
\title{brier_score}
\usage{
brier_score(
  train_test_list,
  match_predict_function = match_predict_xgb(),
  avg = TRUE
)
}
\arguments{
\item{train_test_list}{Output from \code{predict_prepare}}

\item{match_predict_function}{Function to predict treated unit, either  \code{match_predict_xgb} or \code{match_predict_linear}}

\item{avg}{Whether to average (or sum) the briers.}
}
\value{
Result from \code{calc_brier}, length-one double
}
\description{
Takes in training/test data and a prediction function to use, then generates brier score.
}
