% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_score_cv.R
\name{brier_score_cv}
\alias{brier_score_cv}
\title{brier_score_cv}
\usage{
brier_score_cv(
  x_mat,
  match_list,
  design = "cross_all",
  num_folds = 5,
  match_predict_function = match_predict_xgb()
)
}
\arguments{
\item{x_mat}{Typical input matrix}

\item{match_list}{Match result}

\item{design}{See \code{predict_prepare}}

\item{num_folds}{Number of CV folds to use}

\item{match_predict_function}{Function to predict treated units.}
}
\value{
Brier score (averaged over all units)
}
\description{
For a match, calculates brier score using cross validation
}
