% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_score_split.R
\name{brier_score_split}
\alias{brier_score_split}
\title{brier_score_split}
\usage{
brier_score_split(
  x_mat,
  match_list,
  design = "cross_all",
  train_fraction = 0.7,
  match_predict_function = match_predict_xgb()
)
}
\arguments{
\item{x_mat}{Typical input matrix}

\item{match_list}{Match result}

\item{design}{See \code{predict_prepare}}

\item{train_fraction}{Split of data to use for training}

\item{match_predict_function}{Function to predict treated units.}
}
\value{
Brier score
}
\description{
For a match, calculates brier score on a test split of the data.
}
