% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_with_ranks.R
\name{covariance_with_ranks}
\alias{covariance_with_ranks}
\title{covariance_with_ranks}
\usage{
covariance_with_ranks(x_mat, rank_cols = NULL)
}
\arguments{
\item{x_mat}{Matrix of variables (numeric)}

\item{rank_cols}{Names of columns to be converted to ranks before analysis}
}
\value{
Covariance matrix of x_mat (potentially adjusted for ranks)
}
\description{
Computes the covariance of the input matrix, adjusting for ranks if given. If no rank_cols are given, this
function is identical to just calling cov(x_mat)
}
