% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_caliper.R
\name{create_caliper}
\alias{create_caliper}
\title{create_caliper}
\usage{
create_caliper(caliper_list, treat_vec = NULL)
}
\arguments{
\item{caliper_list}{Result of \code{gen_caliper_list}}

\item{treat_vec}{Optional; if you only want pairs between treat and control.}
}
\value{
A matrix, either square (\code{|caliper_vec| x |caliper_vec|})
  or else \code{sum(treat_vec == 1) x sum(treat_vec == 0)}.
}
\description{
Creates caliper penalties.
}
