% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_x_generator.R
\name{default_x_generator}
\alias{default_x_generator}
\title{default_x_generator}
\usage{
default_x_generator(n_rows, n_cols)
}
\arguments{
\item{n_rows}{Number ofrows to produce.}

\item{n_cols}{Number of columns to produce.}
}
\value{
A matrix of data.
}
\description{
Function that generates an input data matrix. Used as default in \code{generate_simulation_input}.
}
