% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_treat_prob_generator.R
\name{example_treat_prob_generator}
\alias{example_treat_prob_generator}
\title{example_treat_prob_generator}
\usage{
example_treat_prob_generator(x_mat)
}
\arguments{
\item{x_mat}{Numeric matrix.}
}
\value{
A vector of probabilities, with length equal to \code{nrow(x_mat)}.
}
\description{
Function that generates treatment probabilities
given an input matrix.
}
