% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{expit}
\usage{
expit(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector, now in (0, 1)
}
\description{
Computes exp / 1 + exp, which is used in several functions.
}
