% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_indexing.R
\name{fold_indexing}
\alias{fold_indexing}
\title{fold_indexing}
\usage{
fold_indexing(index_length, num_folds)
}
\arguments{
\item{index_length}{Length of vector to index}

\item{num_folds}{Number of folds to use.}
}
\value{
List of vectors of indices
}
\description{
Produces a list of indices split into k-folds.
}
