% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_propensity_list.R
\name{gen_propensity_list}
\alias{gen_propensity_list}
\title{gen_propensity_list}
\usage{
gen_propensity_list(
  propensity_function = propensity_score_xgb(),
  oos_propensity = FALSE,
  n_folds = NULL
)
}
\arguments{
\item{propensity_function}{A function that accepts a list with four elements: \code{x_train}, \code{x_test}, \code{y_train}, \code{y_test}, and forms predictions using \code{x_test} (I guess \code{y_test} isn't used yet)}

\item{oos_propensity}{Logical, whether to predict out of sample for the propensity score.}

\item{n_folds}{Default NULL; how many folds if using out of sample propensity.}
}
\value{
Named list, same names as input params.
}
\description{
Takes in elements needed for propensity function, checks input and builds a named list.
}
