% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_p_cut_frame.R
\name{generate_p_cut_frame}
\alias{generate_p_cut_frame}
\title{generate_p_cut_frame}
\usage{
generate_p_cut_frame(
  full_raw_results,
  p_cut_vals = seq(from = 0.2, to = 1, by = 0.05),
  silent = !interactive()
)
}
\arguments{
\item{full_raw_results}{Dataframe of results (spec to come)}

\item{p_cut_vals}{Vector of cut values to plot.}

\item{silent}{Whether to suppress messages as it's running.. Default \code{!interactive()}.}
}
\value{
List per \code{n_rows} value of a dataframe ready to be plotted.
}
\description{
Processes the data into the format required for plotting.
}
