% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_weights.R
\name{generate_random_weights}
\alias{generate_random_weights}
\title{generate_random_weights}
\usage{
generate_random_weights(
  prior_weights,
  number_vectors,
  minimum_weights = NULL,
  hierarchical_list = NULL
)
}
\arguments{
\item{prior_weights}{Must be equal to the length of your columns,
i.e. the length of the weight vectors this function will produce.}

\item{number_vectors}{How many weight vectors to generate.}

\item{minimum_weights}{If you want to set minimums weights.}

\item{hierarchical_list}{List per group / category of variable:
\describe{
  \item{\code{"index"}}{Vector of indices that this group corresponds with}
  \item{\code{"weight"}}{Weight for this group}
  \item{\code{"variance"}}{(Optional) how much variance this group will havve}
}}
}
\value{
list of weight vectors
}
\description{
Generating random weights to use in matching procedures,
each summing to one.
}
