% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_match.R
\name{greedy_match}
\alias{greedy_match}
\title{greedy_match}
\usage{
greedy_match(dist_mat, treat_vec)
}
\arguments{
\item{dist_mat}{Matrix of pairwise distances.}

\item{treat_vec}{Vector representing all subjects; 0 for control,
1 for treated.}
}
\description{
Compute greedy matching without replacement. Alternative to "with replacement" and "optimal".
}
