% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_random_weights.R
\name{hierarchical_random_weights}
\alias{hierarchical_random_weights}
\title{hierarchical_random_weights}
\usage{
hierarchical_random_weights(
  prior_weights,
  number_vectors,
  minimum_weights,
  hierarchical_list
)
}
\arguments{
\item{prior_weights}{Must be equal to the length of your columns,
i.e. the length of the weight vectors this function will produce.}

\item{number_vectors}{How many weight vectors to generate.}

\item{minimum_weights}{If you want to set minimums weights.}

\item{hierarchical_list}{List per group / category of variable:
\describe{
  \item{\code{"index"}}{Vector of indices that this group corresponds with}
  \item{\code{"weight"}}{Weight for this group}
  \item{\code{"variance"}}{(Optional) how much variance this group will havve}
}}
}
\description{
Generates weights in a hierarchical setting
}
