% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_predict_xgb.R
\name{match_predict_xgb}
\alias{match_predict_xgb}
\title{match_predict_xgb}
\usage{
match_predict_xgb(
  nrounds = 50,
  nthread = 1,
  params = list(eta = 0.1, max.depth = 4),
  ...
)
}
\arguments{
\item{nrounds}{Training rounds for the xgb algorithm.}

\item{nthread}{Number of threads to use for fitting, default 1.}

\item{params}{List of params to pass to xgboost, most likely something like \code{eta} and \code{max.depth}}

\item{...}{Additional xgboost params.}
}
\value{
Returns a function that takes in a \code{train_test_list} from \code{predict_prepare}; this function returns a vector of predictions for the test data.
}
\description{
Function to predict treatment / control pairs using xgboost. The returned function takes in training and test data (output from \code{predict_prepare}), trains an xgboost model on the training, predicts on the test, and returns the test vector
}
