% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_propensity_list.R
\name{match_propensity_list}
\alias{match_propensity_list}
\title{match_propensity_list}
\usage{
match_propensity_list(
  propensity_function = propensity_score_xgb(),
  oos_propensity = FALSE,
  n_folds = NULL,
  caliper_sd_mult = 0.6,
  continuous_mult = 100
)
}
\arguments{
\item{propensity_function}{A function that accepts a list with four elements: \code{x_train}, \code{x_test}, \code{y_train}, \code{y_test}, and forms predictions using \code{x_test} (I guess \code{y_test} isn't used yet)}

\item{oos_propensity}{Logical, whether to predict out of sample for the propensity score.}

\item{n_folds}{Default NULL; how many folds if using out of sample propensity.}

\item{caliper_sd_mult}{Maximum gap between units.In \code{sd(propensity_score) * k}, where this parameter is the value k. Default 0.6.}

\item{continuous_mult}{Penality for going above a certain caliper.}
}
\value{
list with names equal to all input params
}
\description{
Generates the propensity parameters used for using propensity-based calipers.
}
