% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_match.R
\name{optimal_match}
\alias{optimal_match}
\title{Computes optimal matches, bipartite}
\usage{
optimal_match(dist_mat, treat_vec, n_sinks = 0, tol_val = 1e-04)
}
\arguments{
\item{dist_mat}{Matrix of pairwise distances.}

\item{treat_vec}{Vector representing all subjects; 0 for control,
1 for treated.}

\item{n_sinks}{single value}

\item{tol_val}{Tolerance for solving optimal matches (how far from the true optimal value is acceptable). Speed with large value, accuracy with small.}
}
\description{
Computes optimal matches, bipartite
}
\keyword{internal}
