% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_brier.R
\name{permutation_brier}
\alias{permutation_brier}
\title{permutation_brier}
\usage{
permutation_brier(
  x_mat,
  match_list,
  design = "cross_all",
  use_cv = TRUE,
  num_permutations = 100L,
  match_predict_function = match_predict_xgb(),
  num_folds = 5,
  train_fraction = 0.7
)
}
\arguments{
\item{x_mat}{Typical input matrix}

\item{match_list}{Match result}

\item{design}{See \code{predict_prepare}}

\item{use_cv}{Logical, default TRUE: wether to use CV to get briers? Else split.}

\item{num_permutations}{Number of permutations to do.}

\item{match_predict_function}{Function to predict treated units.}

\item{num_folds}{If using CV, how many folds to make.}

\item{train_fraction}{If using split, what fraction to train.}
}
\value{
Vector of brier scores for random pairings
}
\description{
For a given match, computes the brier score distribution, assuming the pairings were truly random.
}
