% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results_block.R
\name{plot_results_block}
\alias{plot_results_block}
\title{permutation_matches}
\usage{
plot_results_block(
  n_results,
  x_lim,
  y_lim,
  y_rel = c(0, 1),
  rect_width = 0.006,
  x_shift = 0.015
)
}
\arguments{
\item{n_results}{Number of plots to generate.}

\item{x_lim}{X limits.}

\item{y_lim}{Y limits.}

\item{y_rel}{Y relative size.}

\item{rect_width}{Plot width.}

\item{x_shift}{Horizontal spacing between plots.}
}
\description{
Used in plot_sims to generate the plot specifications.
}
