% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_prepare.R
\name{predict_prepare}
\alias{predict_prepare}
\title{predict_prepare}
\usage{
predict_prepare(
  x_mat,
  index_list,
  design = c("cross_all", "cross_random", "differences_random", "differences_plain")
)
}
\arguments{
\item{x_mat}{Original data, i.e. row for each unit.}

\item{index_list}{List of indices to use in the matching:
\describe{
  \item{\code{treat_train}}{Index of treated units for the training matrix}
  \item{\code{control_train}}{Index of control units for the training matrix}
  \item{\code{treat_test}}{Index of treated units for the test matrix}
  \item{\code{control_test}}{Index of control units for the test matrix}
}}

\item{design}{The design matrix form  to use.}
}
\value{
list:
  \describe{
    \item{\code{x_train}}{Design matrix for training}
    \item{\code{x_test}}{Design matrix for testing}
    \item{\code{y_train}}{Training outcome vector}
    \item{\code{y_test}}{Test outcome vector}
}
}
\description{
Prepares design matrices and other inputs needed for predictive methods
}
