% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_bipartite_matches.R
\name{propensity_bipartite_matches}
\alias{propensity_bipartite_matches}
\title{Propensity match for (default) bipartite}
\usage{
propensity_bipartite_matches(
  x_mat,
  treat_vec,
  match_method = c("with_replacement", "optimal", "greedy"),
  propensity_list = gen_propensity_list(),
  n_sinks = 0,
  caliper_list = gen_caliper_list(),
  sqrt_mahal = TRUE,
  tol_val = NULL
)
}
\arguments{
\item{x_mat}{Input/design matrix}

\item{treat_vec}{Logical (or 1/0) vector, indicating treatment (or control).}

\item{match_method}{"with_replacement", "optimal", or "greedy"}

\item{propensity_list}{See \code{gen_propensity_list}}

\item{n_sinks}{Number of potential matches that don't need to be matched.}

\item{caliper_list}{Forces matches that are close on some metric.}

\item{sqrt_mahal}{Whether to use square root of Mahalanobis distances.}

\item{tol_val}{For optimal matches, you can set a tolerance to be within optimality of, which can be zero for perfect optimality.}
}
\description{
Propensity match for (default) bipartite
}
