% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_score_linear.R
\name{propensity_score_linear}
\alias{propensity_score_linear}
\title{propensity_score_linear}
\usage{
propensity_score_linear(use_linear_lm = FALSE)
}
\arguments{
\item{use_linear_lm}{Whether to use lm or glm.}
}
\value{
returns a function that accepts \code{train_test_list}
and this returns a vector of predictions for the test data
}
\description{
Function to predict treatment using \code{glm} (binomial) or \code{lm}.
}
