% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_score_xgb.R
\name{propensity_score_xgb}
\alias{propensity_score_xgb}
\title{propensity_score_xgb}
\usage{
propensity_score_xgb(
  nrounds = 50,
  nthread = 1,
  params = list(eta = 0.1, max.depth = 3),
  ...
)
}
\arguments{
\item{nrounds}{Training rounds for the xgb algorithm.}

\item{nthread}{Number of threads to use for fitting, default 1.}

\item{params}{List of params to pass to xgboost, most likely something like \code{eta} and \code{max.depth}}

\item{...}{Additional xgboost params.}
}
\value{
returns a function that accepts \code{train_test_list} and this returns a vector of predictions for the test data
}
\description{
Function factory to predict treatment using xgboost.
}
