% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_integer_index.R
\name{rank_integer_index}
\alias{rank_integer_index}
\title{rank_integer_index}
\usage{
rank_integer_index(rank_cols, x_mat)
}
\arguments{
\item{rank_cols}{Integer/number, or logical, or names within \code{colnames(x_mat)}}

\item{x_mat}{the x matrix of interest}
}
\description{
Converts \code{rank_cols} in all allowed forms to an integer index. Takes in ranked cols, either given by integer, logical, or named index. Converts to integer index.
}
