% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranked_x.R
\name{ranked_x}
\alias{ranked_x}
\title{ranked_x}
\usage{
ranked_x(x_mat, rank_cols = NULL, ties_method = "average")
}
\arguments{
\item{x_mat}{Numeric matrix (adjust non-numeric columns prior)}

\item{rank_cols}{Names or index of columns to be converted to ranks}

\item{ties_method}{How to break ties in ranks, by default uses "average" (same as the rank function's default).
for further options}
}
\value{
x_mat, potentially adjusted for ranks
}
\description{
Converts indicated columns to ranked versions of themselves. Takes a numeric matrix and converts any columns indicated by the \code{rank_cols} input to their ranks.
}
