% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_from_one_func.R
\name{rmse_from_one_func}
\alias{rmse_from_one_func}
\title{rmse_from_one_func}
\usage{
rmse_from_one_func(vec)
}
\arguments{
\item{vec}{Vector of estimates around 1.}
}
\value{
Returns a single number: the RMSE.
}
\description{
Calculates RMSE with a known target of one.
}
