% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_sink_wrap.R
\name{simple_sink_wrap}
\alias{simple_sink_wrap}
\title{simple_sink_wrap}
\usage{
simple_sink_wrap(simple_match_list, n_sinks = NULL, weight_vec = NULL)
}
\arguments{
\item{simple_match_list}{match result from one of \code{with_replacement_match}, \code{greedy_match}, \code{with_replacement_nbp_match} or \code{greedy_nbp_match}.}

\item{n_sinks}{default NULL, vector of sink values to use.}

\item{weight_vec}{Default \code{NULL}: optionally supply the weight vector used to generate \code{dist_mat} and it will be returned in the \code{match_list} generated from this function}
}
\value{
List of lists.
}
\description{
Wraps a match_list from a simple match result and removes various numbers of sinks.
}
