% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_mahal.R
\name{weighted_mahal}
\alias{weighted_mahal}
\title{weighted_mahal}
\usage{
weighted_mahal(
  x_mat,
  cov_x,
  weight_vec = NULL,
  treat_vec = NULL,
  sqrt_mahal = TRUE,
  partial_index = NULL
)
}
\arguments{
\item{x_mat}{Numeric matrix (adjust non-numeric columns prior), already rank-adjusted if desired.}

\item{cov_x}{Covariance of x, calculated potentially with ranks.}

\item{weight_vec}{Vector of weights corresponding to columns of x_mat, giving weights relative to normal Mahalanobis.}

\item{treat_vec}{optionally specify which units are treated.}

\item{sqrt_mahal}{Logical, default TRUE; whether to use regular Mahalanobis.}

\item{partial_index}{How to partition the full matrix.}
}
\value{
returns a matrix of pairwise distances; the relevant indexing depends on \code{treat_vec} and \code{partial_index}
}
\description{
Computes weighted Mahalanobis distance, using Choleski decomp.
}
