% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_replacement_match.R
\name{with_replacement_match}
\alias{with_replacement_match}
\title{with_replacement_match}
\usage{
with_replacement_match(dist_mat, treat_vec)
}
\arguments{
\item{dist_mat}{Matrix of pairwise distances.}

\item{treat_vec}{Vector representing all subjects; 0 for control,
1 for treated.}
}
\description{
Computes matching with replacement. Alternative to "greedy" and "optimal".
}
