% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_plots.R
\name{generate_p_cut_frame}
\alias{generate_p_cut_frame}
\title{Processes the data into the format required for plotting.}
\usage{
generate_p_cut_frame(
  full_raw_results,
  p_cut_vals = seq(from = 0.2, to = 1, by = 0.05),
  silent = !interactive()
)
}
\arguments{
\item{full_raw_results}{Dataframe of results (spec to come)}

\item{p_cut_vals}{Vector of cut values to plot.
\code{rmse_from_one_func}, which gives RMSE relative to one.}

\item{silent}{***}
}
\value{
List per \code{n_rows} value of a dataframe ready to be plotted.
}
\description{
Processes the data into the format required for plotting.
}
